+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.4.0]
+## [0.4.1] - 2020-12-21
+### Fixed
+- Wrong **TransformQueue--** action.
+
+## [0.4.0] - 2020-12-20
 ### Changed
 - Create multiple connections for Sender and send a number of packs to each before close.